feat: Add cropped_input.py for video cropping
This commit adds a new file, cropped_input.py, to the preprocessing/utils directory. The cropped_input.py file contains functions for cropping frames in videos and saving them in a new folder. The main function, crop_videos, iterates over all video files in a given folder, crops the frames, and saves them in the "data/cropped" folder. The crop_frames function masks the frames in a video and saves them in the specified output folder. The get_video_name function extracts the name of the video from the video path. The main function in cropped_input.py is used to demonstrate the usage of the crop_videos function.